requirements.txt
asgiref==3.7.2
colorama==0.4.6
Django==5.0
djangorestframework==3.14.0
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
psycopg2==2.9.9
pytest==7.4.3
pytest-django==4.7.0
python-decouple==3.8
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
